requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
gymnasium>=0.29.0  # 使用新版本的 Gymnasium 替代 gym
pyyaml>=5.4.1
scikit-learn>=1.0.0
matplotlib>=3.4.2
seaborn>=0.11.2
jupyter>=1.0.0
pytest>=6.2.5